// data/filterStore/enums.ts
export enum RoleFilter {
  tank = 'tank',
  apk = 'apk',
  adk = 'adk',
  support = 'support',
  fighter = 'fighter',
  assassin = 'assassin',
}
export enum PositionFilter {
  top = 'top',
  mid = 'mid',
  bot = 'bot',
  jungle = 'jungle',
}

export enum RungFilter {
  max = 'max',
  high = 'high',
  low = 'low',
}